Appropriately size buffer when reading footer
Using the default buffer size (4K) when reading the footer causes
repeated small reads rather than one read of the correct size. This is
especially bad when using a datasource like S3 that charges a cost
per request regardless of the size of data returned.

Ideally ReadFooter could estimate the footer size and attempt to read
the footer size and the footer in a single read operation, only making a
second request if it under-estimated, but doing that correctly is much
harder.