Introduce SeekRowGroup API
Add a new method, SeekRowGroup, to the ParquetReader that logically
moves the reader to the start of the indexed row group, such that the
next read will start by returning the first row of that row group.

This method does no IO on the underlying file, such that clients can
skip over row groups that contain records that they don't need to read
without needing to do work proportional to the size of the file.