Add MarshalFast implementation to improve writer performance
Provide an alternative Marshal implementation, MarshalFast, that
minimizes use of reflect to maximize performance and minimize
allocations.

The Marshal implementation spends a huge amount of time doing
allocations (and subsequently garbage collection), largely due to
using reflection on every element. Instead, we can use reflection to
analyze elements once-per-type and construct an encoder for elements of
that type. Then, the encoder can walk the pointers of each element of
that type and store its root values in the tables.

MarshalFast liberally uses the unsafe package, because it walks unsafe
pointers, and adds a dependency on goccy/go-reflect for some type
analysis features.

MarshalFast supports (as far as I can tell) all use-cases of Marshal
except for map types, which are hard to walk without using reflection.