support nesting in map struct
previously missing keys with children were ignored in MapStruct marshal.
this leads to missing values, and the code eventually panics in -
chunkMap[name] = layout.PagesToChunk(pages)
since pages is an empty slices in this case, and PagesToChunk assumes it
has at least one element.

the fix is to push to the stack all the missing keys in the map -
- simple fields (without children) get a null value (as was done
  previously)
- nested fields (with children) get an empty map value